bug/wrong PR name is generated
