+++ internal/helper/consts.go
@@ -3,4 +3,6 @@ package helper
 const (
 	pushYes  = "y"
 	oneSpace = " "
+	maxDevBranchName = 50
+	msymbol          = "-"
 )
